% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sel_Index.R
\name{AugEI_fun}
\alias{AugEI_fun}
\title{AugEI_fun}
\usage{
AugEI_fun(CV_gpnumber, candi, gp1, n, num1, num2, kinship, p_sim)
}
\arguments{
\item{CV_gpnumber}{number of folds for cross validation}

\item{candi}{number vector; length is equal to total candidates}

\item{gp1}{the number of group for the size num1}

\item{n}{the total number of candidates}

\item{num1}{num1+num2=n}

\item{num2}{num1+num2=n}

\item{kinship}{kinship matrix}

\item{p_sim}{phenotype data for genotypes (calculate by GBLUP model)}
}
\description{
Function deal with GBLUP model and calculate AugEI value for each genotype
This can go on for multiple lines.
}
