% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugEITD-package.R
\docType{package}
\name{AugEITD-package}
\alias{AugEITD}
\alias{AugEITD-package}
\title{AugEITD: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hui-Ning Tu \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
