% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMethods.R
\name{EstimationM}
\alias{EstimationM}
\title{EstimationM}
\usage{
EstimationM(
  kin,
  snp_matrix,
  opt_trainSet,
  pheno,
  delta = c(1/5, 1/4, 1/3, 1/2, 2/3, 3/4, 1),
  n = nrow(kin)
)
}
\arguments{
\item{kin}{kinship matrix}

\item{snp_matrix}{the SNP marker matrix that has been organized (finishing quality control, imputation and standardized)}

\item{opt_trainSet}{the number of group for the size num1}

\item{pheno}{collected phenotype data; please include species ID and the column names; if data have population structure, please include subpopulations information in the dataframe}

\item{delta}{training set size = delta *n; default:1/5,1/4,1/3,1/2,2/3,3/4,1}

\item{n}{the number of genotypes in the dataset}
}
\description{
compare bayesian RKHS (GBLUP model), RKHS, random forest and McRank
}
