% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sel_Index.R
\name{GenerateTS}
\alias{GenerateTS}
\title{GenerateTS}
\usage{
GenerateTS(
  kinship,
  nOpsim = 1000,
  nEvalsim = 2000,
  CV_gpnumber = 5,
  h = 0.5,
  n,
  sg = 25,
  mu = 100,
  subpopTag,
  desireH = c(0.5),
  desireDelta = c(1/5, 1/3, 2/3)
)
}
\arguments{
\item{kinship}{kinship matrix}

\item{nOpsim}{Number of repetitions for cross-validation}

\item{nEvalsim}{Number of repetitions for evaluate the performance of optimal training set}

\item{CV_gpnumber}{number of folds for cross validation}

\item{h}{heritability parameters for obtaining the optimal training set; default is 0.5}

\item{n}{the total number of candidates}

\item{sg}{variance of genotypic values; default is 25}

\item{mu}{general mean; default is 100}

\item{subpopTag}{Information of population structure. Input dataset should include genotypes' name and subpopulation tags. If the dataset don not consider the population structure, the parameters is allowed to miss.}

\item{desireH}{The heratibility parameters for user's desiring traits. This would be used in the simulation step.}

\item{desireDelta}{The proportion of the candidate dataset that users would like to use as the training set.}
}
\description{
Main function.This function will generate the optimal training set for the given dataset.
Users can input the desiring sample size and heritability to obtain the NDCG and relative NDCG results based on the settings and the optimal training set.
The index of the suggested training set will also be returned.
}
