% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-genorice44K.R
\docType{data}
\name{geno_rice44K}
\alias{geno_rice44K}
\title{A genomic relationship data (kinship matrix) of 44K rice (data with strong population structure).}
\format{
A matrix with 301 rows and columns:
The row names and column names referred to the name of genotypes.
}
\source{
\url{https://doi.org/10.1038/ncomms1467}
}
\usage{
geno_rice44K
}
\description{
The kinship matrix is calculated by the function: K = X\%*\%t(X)/ncol(X), where X stands for standardized SNP marker matrix.
The original dataset included 413 genotypes, but we used only 301 genotypes.
Report ...
}
\keyword{datasets}
