% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-genotrop.R
\docType{data}
\name{geno_trop}
\alias{geno_trop}
\title{A genomic relationship data (kinship matrix) of tropical rice}
\format{
A matrix with 328 rows and columns:
The row names and column names referred to the name of genotypes.
}
\source{
\url{https://doi.org/10.1371/journal.pgen.1004982}
}
\usage{
geno_trop
}
\description{
A kinship matrix of tropical rice (data without population structure).
This kinship matrix is calculated by arranged SNP marker matrix.
Report ...
}
\keyword{datasets}
