% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-subpop_tag.R
\docType{data}
\name{subpop_tag}
\alias{subpop_tag}
\title{A data frame with subpopulations information of rice 44K.}
\format{
A data frame with 413 rows and 2 variables:
\describe{
\item{NSFTV.ID}{ID of the genotypes}
\item{Sub.population}{Subpopulation of genotypes}
}
}
\source{
\url{https://doi.org/10.1038/ncomms1467}
}
\usage{
subpop_tag
}
\description{
To consider population structure when obtaining the optimal training set, we need the subpopulation information of genotypes.
Report ...
}
\keyword{datasets}
