% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugEITD-package.R
\docType{package}
\name{TSOFIBG-package}
\alias{TSOFIBG}
\alias{TSOFIBG-package}
\title{TSOFIBG: Training set determination for genomic selection by using bayesian optimization}
\description{
Determine the training set based on user's genotype data of plants. Functions return the selection index of each genotype and the efficiency of the training set under different sizes.
}
\author{
\strong{Maintainer}: Hui-Ning Tu \email{huining0312@gmail.com}

Authors:
\itemize{
  \item Chen-Tuo Liao \email{ctliao@ntu.edu.tw} (\href{https://orcid.org/0000-0001-9777-3701}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
